chore: rename utils to shared
