+++ mission3/src/article/commands/get-article/get-article.controller.js
@@ -1,6 +1,6 @@
 import { getArticleService } from './get-article.service.js';
-import { AppError } from '../../../utils/AppError.js';
-import { asyncHandler } from '../../../utils/asyncHandler.js';
+import { AppError } from '../../../shared/AppError.js';
+import { asyncHandler } from '../../../shared/asyncHandler.js';
 
 export const getArticleController = asyncHandler(async (req, res) => {
   const { articleId } = req.validatedParams;
