+++ mission3/src/comment/commands/create-comment/create-comment.controller.js
@@ -1,6 +1,6 @@
 // create-comment.controller.js
 import { createCommentService } from './index.js';
-import { asyncHandler } from '../../../utils/index.js';
+import { asyncHandler } from '../../../shared/index.js';
 
 export const createCommentController = asyncHandler(async (req, res) => {
   const created_comment = await createCommentService(req.validatedBody);
