+++ mission3/src/comment/commands/delete-comment/delete-comment.controller.js
@@ -1,6 +1,6 @@
 import { deleteCommentService } from './index.js';
-import { AppError } from '../../../utils/index.js';
-import { asyncHandler } from '../../../utils/index.js';
+import { AppError } from '../../../shared/index.js';
+import { asyncHandler } from '../../../shared/index.js';
 
 export const deleteCommentController = asyncHandler(async (req, res) => {
   const { commentId } = req.validatedParams;
