+++ mission3/src/comment/commands/get-comment/get-comment.controller.js
@@ -1,6 +1,6 @@
 import { getCommentService } from './get-comment.service.js';
-import { AppError } from '../../../utils/AppError.js';
-import { asyncHandler } from '../../../utils/asyncHandler.js';
+import { AppError } from '../../../shared/AppError.js';
+import { asyncHandler } from '../../../shared/asyncHandler.js';
 
 export const getCommentController = asyncHandler(async (req, res) => {
   const { commentId } = req.validatedParams;
