+++ mission3/src/comment/commands/list-comments/list-product-comments.controller.js
@@ -1,4 +1,4 @@
-import { asyncHandler } from '../../../utils/asyncHandler.js';
+import { asyncHandler } from '../../../shared/asyncHandler.js';
 import { listCommentsService } from './list-comments.service.js';
 
 export const listProductCommentsController = asyncHandler(async (req, res) => {
