+++ mission3/src/product/commands/create-product/create-product.controller.js
@@ -1,6 +1,6 @@
 // create-product.controller.js
 import { createProductService } from './index.js';
-import { asyncHandler } from '../../../utils/index.js';
+import { asyncHandler } from '../../../shared/index.js';
 
 export const createProductController = asyncHandler(async (req, res) => {
   const created_product = await createProductService(req.validatedBody);
