+++ mission3/src/product/commands/list-products/list-products.controller.js
@@ -1,4 +1,4 @@
-import { asyncHandler } from '../../../utils/asyncHandler.js';
+import { asyncHandler } from '../../../shared/asyncHandler.js';
 import { listProductsService } from './list-products.service.js';
 
 export const listProductsController = asyncHandler(async (req, res) => {
