+++ mission3/src/shared/asyncHandler.js
@@ -1,3 +1,3 @@
-// src/utils/asyncHandler.js
+// src/shared/asyncHandler.js
 export const asyncHandler = (fn) => (req, res, next) =>
   Promise.resolve(fn(req, res, next)).catch(next);
